Actividad03 finalizada

+++ Actividad03/Actividad03/actividad03/introduceDatos/Pregunta.java
@@ -4,11 +4,16 @@ import java.io.BufferedReader;
 import java.io.InputStreamReader;
 
 public class Pregunta {
+    /**
+     * Esta clase contiene una serie de metodos que reciben un input de usuario y lo transforman en un valor númerico haciendo
+     * las correspondientes validaciones para obtener un output valido
+     */
 
     static BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
     static String inputUsuario= "";
 
     public static int pideEntero(String pregunta){
+        //Convierte el input del usuario en un integer
         int entero= 0;
         boolean preguntar= true;
         System.out.println(pregunta);
@@ -30,6 +35,7 @@ public class Pregunta {
     }
 
     public static double pideDouble(String pregunta){
+        //Convierte el input del usuario en un double
         double numDecimal= 0.0;
         boolean preguntar= true;
         System.out.println(pregunta);
@@ -37,7 +43,7 @@ public class Pregunta {
         do{
             try {
                 inputUsuario = br.readLine();
-                numDecimal = Integer.parseInt(inputUsuario);
+                numDecimal = Double.parseDouble(inputUsuario);
                 preguntar= false;
             } catch (Exception e) {
                 System.out.println("Introduce un numero entero sin decimales de 0 a X");
